requirements.txt
matplotlib
Pillow
torch
torchvision
